+++ src/Service/JiraApiService.php
@@ -23,14 +23,16 @@ use App\Model\Planning\SprintSum;
 use App\Model\SprintReport\SprintReportData;
 use App\Model\SprintReport\SprintReportEpic;
 use App\Model\SprintReport\SprintReportIssue;
+use App\Model\SprintReport\SprintReportProject;
+use App\Model\SprintReport\SprintReportProjects;
 use App\Model\SprintReport\SprintReportSprint;
+use App\Model\SprintReport\SprintReportVersions;
 use App\Model\SprintReport\SprintStateEnum;
 use Doctrine\Common\Collections\ArrayCollection;
 use Symfony\Contracts\HttpClient\HttpClientInterface;
 
-class JiraApiService implements ApiServiceInterface, DataProviderServiceInterface
+class JiraApiService implements DataProviderServiceInterface
 {
-    private const PROJECT_TRACKER_IDENTIFIER = 'JIRA';
     private const CPB_ACCOUNT_MANAGER = 'anbjv';
     private const NO_SPRINT = 'NoSprint';
     private const API_PATH_SEARCH = '/rest/api/2/search';
@@ -67,77 +69,51 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         ];
     }
 
-    public function getProjectTrackerIdentifier(): string
-    {
-        return self::PROJECT_TRACKER_IDENTIFIER;
-    }
-
-    /**
-     * @throws ApiServiceException
-     */
     public function getAllProjectCategories(): mixed
     {
         return $this->get(self::API_PATH_PROJECT_CATEGORIES);
     }
 
-    /**
-     * Get all accounts.
-     *
-     * @throws ApiServiceException
-     */
     public function getAllAccounts(): mixed
     {
         return $this->get(self::API_PATH_ACCOUNT);
     }
 
-    /**
-     * Get all accounts.
-     *
-     * @throws ApiServiceException
-     */
     public function getAllCustomers(): mixed
     {
         return $this->get(self::API_PATH_CUSTOMERS);
     }
 
-    /**
-     * Get all projects, including archived.
-     *
-     * @throws ApiServiceException
-     */
     public function getAllProjects(): mixed
     {
         return $this->get(self::API_PATH_PROJECT);
     }
 
-    /**
-     * Get project.
-     *
-     * @param $key
-     *   A project key or id
-     *
-     * @throws ApiServiceException
-     */
     public function getProject($key): mixed
     {
         return $this->get(self::API_PATH_PROJECT_BY_ID.$key);
     }
 
-    /**
-     * Get current user permissions.
-     *
-     * @throws ApiServiceException
-     */
     public function getCurrentUserPermissions(): mixed
     {
         return $this->get(self::API_PATH_MY_PERMISSIONS);
     }
 
-    /**
-     * Get list of allowed permissions for current user.
-     *
-     * @throws ApiServiceException
-     */
+    public function getSprintReportProjects(): SprintReportProjects
+    {
+        throw new ApiServiceException('Method not implemented', 501);
+    }
+
+    public function getSprintReportProject(string $projectId): SprintReportProject
+    {
+        throw new ApiServiceException('Method not implemented', 501);
+    }
+
+    public function getSprintReportProjectVersions(string $projectId): SprintReportVersions
+    {
+        throw new ApiServiceException('Method not implemented', 501);
+    }
+
     public function getPermissionsList(): array
     {
         $list = [];
@@ -153,15 +129,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $list;
     }
 
-    /**
-     * Create a jira project.
-     *
-     * See https://docs.atlassian.com/software/jira/docs/api/REST/9.3.0/#api/2/project-createProject
-     *
-     * @return ?string
-     *
-     * @throws ApiServiceException
-     */
     public function createProject(array $data): ?string
     {
         $projectKey = strtoupper($data['form']['project_key']);
@@ -185,11 +152,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $response->key == $projectKey ? $projectKey : null;
     }
 
-    /**
-     * Create a jira customer.
-     *
-     * @throws ApiServiceException
-     */
     public function createTimeTrackerCustomer(string $name, string $key): mixed
     {
         return $this->post(self::API_PATH_CUSTOMERS,
@@ -201,11 +163,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         );
     }
 
-    /**
-     * Create a Jira account.
-     *
-     * @throws ApiServiceException
-     */
     public function createTimeTrackerAccount(string $name, string $key, string $customerKey, string $contactUsername): mixed
     {
         return $this->post(self::API_PATH_ACCOUNT,
@@ -229,26 +186,11 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         );
     }
 
-    /**
-     * Get tempo account base on key.
-     *
-     * @throws ApiServiceException
-     */
     public function getTimeTrackerAccount(string $key): mixed
     {
         return $this->get(self::API_PATH_ACCOUNT_BY_KEY.$key);
     }
 
-    /**
-     * Create a project link to account.
-     *
-     * @param mixed $project
-     *                       The project that was created on form submit
-     * @param mixed $account
-     *                       The account that was created on form submit
-     *
-     * @throws ApiServiceException
-     */
     public function addProjectToTimeTrackerAccount(mixed $project, mixed $account): void
     {
         $this->post(self::API_PATH_LINK_PROJECT_TO_ACCOUNT, [
@@ -261,11 +203,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         ]);
     }
 
-    /**
-     * Create project board.
-     *
-     * @throws ApiServiceException
-     */
     public function createProjectBoard(string $type, mixed $project): void
     {
         // If no template is configured don't create a board.
@@ -298,19 +235,11 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         ]);
     }
 
-    /**
-     * Get account based on id.
-     *
-     * @throws ApiServiceException
-     */
     public function getAccount(string $accountId): mixed
     {
         return $this->get(self::API_PATH_ACCOUNT.$accountId.'/');
     }
 
-    /**
-     * @throws ApiServiceException
-     */
     public function getRateTableByAccount(string $accountId): mixed
     {
         return $this->get(self::API_PATH_RATE_TABLE, [
@@ -319,9 +248,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         ]);
     }
 
-    /**
-     * @throws ApiServiceException
-     */
     public function getAccountIdsByProject(string $projectId): array
     {
         $projectLinks = $this->get(self::API_PATH_ACCOUNT_IDS_BY_PROJECT.$projectId);
@@ -333,24 +259,11 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         }, []);
     }
 
-    /**
-     * Get all boards.
-     *
-     * @throws ApiServiceException
-     */
     public function getAllBoards(): mixed
     {
         return $this->get(self::API_PATH_BOARD);
     }
 
-    /**
-     * Get all sprints for a given board.
-     *
-     * @param string $boardId board id
-     * @param string $state sprint state. Defaults to future,active sprints.
-     *
-     * @throws ApiServiceException
-     */
     public function getAllSprints(string $boardId, string $state = 'future,active'): array
     {
         $sprints = [];
@@ -374,16 +287,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $sprints;
     }
 
-    /**
-     * Get all issues for given board and sprint.
-     *
-     * @param string $boardId id of the jira board to extract issues from
-     * @param string $sprintId id of the sprint to extract issues for
-     *
-     * @return array array of issues
-     *
-     * @throws ApiServiceException
-     */
     public function getIssuesInSprint(string $boardId, string $sprintId): array
     {
         $issues = [];
@@ -416,12 +319,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $issues;
     }
 
-    /**
-     * Create data for planning page.
-     *
-     * @throws ApiServiceException
-     * @throws \Exception
-     */
     public function getPlanningData(): PlanningData
     {
         $planning = new PlanningData();
@@ -722,10 +619,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         throw new ApiServiceException('Sprint not found', 404);
     }
 
-    /**
-     * @throws ApiServiceException
-     * @throws \Exception
-     */
     public function getSprintReportData(string $projectId, string $versionId): SprintReportData
     {
         $sprintReportData = new SprintReportData();
@@ -1041,9 +934,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return false;
     }
 
-    /**
-     * @throws ApiServiceException
-     */
     public function getClientDataForProject(string $projectId): array
     {
         $clients = [];
@@ -1090,9 +980,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $clients;
     }
 
-    /**
-     * @throws ApiServiceException
-     */
     public function getAllProjectData(): array
     {
         $projects = [];
@@ -1140,10 +1027,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         ]);
     }
 
-    /**
-     * @throws ApiServiceException
-     * @throws \Exception
-     */
     public function getWorklogDataForProject(string $projectId): array
     {
         $worklogsResult = [];
@@ -1170,9 +1053,6 @@ class JiraApiService implements ApiServiceInterface, DataProviderServiceInterfac
         return $worklogsResult;
     }
 
-    /**
-     * @throws ApiServiceException
-     */
     public function getAllAccountData(): array
     {
         $accountsResult = [];