+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Modified getSprintReportData to work with Leantime data
 * Added project lead to client when syncing projects.
 * Remove description from create invoice page.
 * Added generate description button to invoice when client is set.